--- test/rtsp_audio/test_webhook.py
@@ -1,12 +1,0 @@
-from flask import Flask, request
-
-app = Flask(__name__)
-
-@app.route("/audio", methods=["POST"])
-def receive():
-    print("=== GOT POST ===")
-    print(request.json)
-    print("================")
-    return "ok", 200
-
-app.run(host="0.0.0.0", port=8080)
